# aulas/praticas/pratica09/swagger.yaml
openapi: 3.0.0
info:
  title: API de Produtos José Lucas - Prática 9
  description: API simples para gerenciar produtos (exemplo para prática).
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Ambiente de desenvolvimento
tags:
  - name: produtos
paths:
  /produtos:
    get:
      tags:
        - produtos
      summary: Lista todos os produtos
      responses:
        '200':
          description: Uma lista de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayOfProdutos'
    post:
      tags:
        - produtos
      summary: Cria um novo produto
      requestBody:
        description: Um JSON com dados do novo produto
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '422':
          description: Nome e preço são obrigatórios

  /produtos/{produtoId}:
    get:
      tags:
        - produtos
      summary: Busca um produto por id
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      responses:
        '200':
          description: Um JSON com os dados do produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado

    put:
      tags:
        - produtos
      summary: Atualiza um produto
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      requestBody:
        description: Um JSON com dados do produto a atualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        '200':
          description: Um JSON com os dados do produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado
        '422':
          description: Nome e preço são obrigatórios

    delete:
      tags:
        - produtos
      summary: Remove um produto
      parameters:
        - in: path
          name: produtoId
          required: true
          description: Id do produto a ser buscado
          schema:
            type: integer
      responses:
        '204':
          description: Produto removido com sucesso
        '404':
          description: Produto não encontrado

components:
  schemas:
    ArrayOfProdutos:
      type: array
      items:
        $ref: '#/components/schemas/Produto'
    Produto:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        preco:
          type: number
      required:
        - nome
        - preco